requirements.txt
alexflipnote.py
cogwatch
dblpy
statcord.py
mysql.connector
timeago
jishaku
speedtest-cli
discord_argparse
argparse
blistpy
ibl
asyncpraw
asyncio
discord.py[voice]
matplotlib
googletrans
nekos.py
buttons
cairosvg
discord-ext-menus
asyncpg
aiocron
wavelink
parsedatetime
discordlists.py
akanekopy
topggpy
ujson
colorama
colorlog
aiomysql
asyncmy
psycopg[binary,pool]
